Some fixes, tweaks, and additions. - Existence of requests will be checked before they're used (TODO: check their validity, especially if they somehow interact with the system). - Changed column widths of the front page. - Added price to the front page's table. - Added some USA specific items, e.g. dollar sign and date format (TODO: make this configurable, or better, go off of server localle settings). - Made the html root the "htdocs" folder (instead of the home inventory root directory, because users ought to not have direct access to the templates, etc).

+++ htdocs/editPicture.php
@@ -8,7 +8,7 @@
 
 	
 	// save pictures
-	if ($_REQUEST['action'] == 'save')
+	if (isset($_REQUEST['action']) && $_REQUEST['action'] == 'save')
 	{
 		for ($i = 1; $i <= 4; $i++)
 		{
@@ -26,7 +26,7 @@
 			$result = mysql_query($sSql) or die('Query failed: ' . mysql_error());
 		}
 	}
-	elseif ($_REQUEST['action'] == 'deletePicture' && is_numeric($_REQUEST['pictId']))
+	elseif (isset($_REQUEST['action']) && $_REQUEST['action'] == 'deletePicture' && is_numeric($_REQUEST['pictId']))
 	{
 		$sSql = "DELETE FROM Picture WHERE PCT_ID = {$_REQUEST['pictId']}";
 		$result = mysql_query($sSql) or die('Query failed: ' . mysql_error());
